requirements.txt
requests
selenium
bs4
colorama
pycryptodome
pypiwin32
PIL-Tools
pylibcheck
pyinstaller
psutil
keyboard
httpx
pipwin
pyotp
pycryptodome
zstandard
Nuitka
ordered-set
